magit-push-{,to-remote-}dwim: new commands
Add new commands `magit-push-dwim' and `magit-push-to-remote-dwim'.
These commands push to some remote without using an explicit refspec.
What is being pushed instead depends on various Git variables.

`magit-push-dwim' is identical to `magit-push-implicitly', which was
added some time after v2.1, and which was later removed again because I
concluded it was unnecessary given recent changes.

But in the spirit of the slogan for the upcoming release, "deal with the
fact that everyone has a different opinion on how branching and pushing
should work", I am making an equivalent command available again.

But I am not adding it to the popup by default (that would be confusing),
to do that you have to add something like this to you init file:

  (with-eval-after-load 'magit-remote
    (magit-define-popup-action 'magit-push-popup ?P
      'magit-push-dwim--desc
      'magit-push-dwim ?p t))

Note that instead of a constant string, the function
`magit-push-dwim--desc', or more precisely, its return value is used as
the action description.  That function consults various Git variables to
determine what is going to happen when running `git push' without
explicitly specifying the refspec (like the other push variants do).

There are many variables involved in this and I am not 100% sure that
this function takes each one of them into account they way it should, so
the displayed information might not be correct.

`magit-push-to-remote-dwim' is a completely new command.  Like
`magit-push-dwim' it does not use an explicit refspec, it does however
prompt the user for the remote to push to.

This command too isn't available from the popup by default.  To add it
use something like:

  (with-eval-after-load 'magit-remote
    (magit-define-popup-action 'magit-push-popup ?r
      'magit-push-to-remote-dwim--desc
      'magit-push-to-remote-dwim ?p t)